+++ sdk/python/packages/flet-ads/src/flet_ads/consent_manager.py
@@ -29,12 +29,12 @@ class ConsentManager(ft.Service):
     More info in the [AdMob documentation](https://developers.google.com/admob/flutter/privacy).
 
     Raises:
-        FletUnsupportedPlatformException: When any of its methods are called on
-            a web and/or non-mobile platform.
+        FletUnsupportedPlatformException: When used on a web and/or non-mobile
+            platform.
     """  # noqa: E501
 
-    def init(self):
-        super().init()
+    def before_update(self):
+        super().before_update()
         if self.page.web or not self.page.platform.is_mobile():
             raise ft.FletUnsupportedPlatformException(
                 f"{self.__class__.__name__} is only supported on "